backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
beautifulsoup4
requests
python-dotenv
youtube-transcript-api
pytube
